+++ src/cpp/src/whisper/whisper_utils.hpp
@@ -24,6 +24,11 @@ ov::genai::WhisperGenerationConfig prepare_per_generate_config(
     const ov::genai::WhisperGenerationConfig& base_config,
     const ov::genai::OptionalWhisperGenerationConfig& per_generate_config);
 
+std::string find_language_by_token_id(const std::map<std::string, int64_t>& lang_to_id, int64_t token_id);
+
+// "<|en|>" -> "en"
+std::string to_unescaped_language(const std::string& language);
+
 }  // namespace utils
 }  // namespace genai
 }  // namespace ov
